+++ modules/tf-azure-bastion/main.tf
@@ -3,20 +3,7 @@
  *
  * This module installs an Azure Bastion Host and its public IP address.
  *
- * This module also configures the 'AzureBastionSubnet' in the provided Virtual Network
- * with the provided subnet prefix.
  */
-
-# Subnet
-resource "azurerm_subnet" "snet" {
-  count = var.sku != "Developer" ? 1 : 0
-
-  name                            = "AzureBastionSubnet"
-  resource_group_name             = var.resource_group_name
-  virtual_network_name            = regex(".*/virtualNetworks/(.*)", var.virtual_network_id)[0]
-  address_prefixes                = [var.subnet_prefix]
-  default_outbound_access_enabled = false
-}
 
 # Public IP
 resource "azurerm_public_ip" "pip" {
@@ -48,7 +35,7 @@ resource "azurerm_bastion_host" "bastion" {
 
     content {
       name                 = "ipconf"
-      subnet_id            = azurerm_subnet.snet[0].id
+      subnet_id            = var.subnet_id
       public_ip_address_id = var.new_public_ip_address != null ? azurerm_public_ip.pip[0].id : var.public_ip_address_id != null ? var.public_ip_address_id : null
     }
   }