+++ modules/tf-azure-vpn/main.tf
@@ -3,8 +3,6 @@
  *
  * This module installs an Azure VPN Connection.
  *
- * This module also configures the 'GatewaySubnet' in the provided Virtual Network
- * with the provided subnet prefix.
  */
 
 # Public IPs
@@ -39,7 +37,7 @@ resource "azurerm_virtual_network_gateway" "vnetgw" {
   enable_bgp    = var.virtual_network_gateway.bgp != null ? true : false
   sku           = var.virtual_network_gateway.sku
 
-  dynamic ip_configuration {
+  dynamic "ip_configuration" {
     for_each = var.virtual_network_gateway.instances
 
     content {
@@ -50,7 +48,7 @@ resource "azurerm_virtual_network_gateway" "vnetgw" {
     }
   }
 
-  dynamic custom_route {
+  dynamic "custom_route" {
     for_each = var.virtual_network_gateway.custom_route != null ? [true] : []
 
     content {
@@ -62,11 +60,11 @@ resource "azurerm_virtual_network_gateway" "vnetgw" {
     for_each = var.virtual_network_gateway.bgp != null ? var.virtual_network_gateway.instances : {}
 
     content {
-      asn = var.virtual_network_gateway.bgp.asn
+      asn         = var.virtual_network_gateway.bgp.asn
       peer_weight = var.virtual_network_gateway.bgp.peer_weight
       peering_addresses {
         ip_configuration_name = bgp_settings.key
-        apipa_addresses = bgp_settings.value.bgp_apipa_addresses
+        apipa_addresses       = bgp_settings.value.bgp_apipa_addresses
       }
     }
   }
@@ -74,7 +72,7 @@ resource "azurerm_virtual_network_gateway" "vnetgw" {
 
 # Local Network Gateways
 resource "azurerm_local_network_gateway" "lnetgw" {
-  for_each = { for lgw in var.local_network_gateways: lgw.name => lgw }
+  for_each = { for lgw in var.local_network_gateways : lgw.name => lgw }
 
   name                = each.value.name
   location            = var.location
@@ -88,48 +86,48 @@ resource "azurerm_local_network_gateway" "lnetgw" {
 
     content {
       bgp_peering_address = bgp_settings.key
-      asn = bgp_settings.value.asn
-      peer_weight = bgp_settings.value.peer_weight
+      asn                 = bgp_settings.value.asn
+      peer_weight         = bgp_settings.value.peer_weight
     }
   }
 }
 
 # Tunnel Connections
 resource "azurerm_virtual_network_gateway_connection" "connection" {
-  for_each = { for vcn in var.connections: vcn.name => vcn }
+  for_each = { for vcn in var.connections : vcn.name => vcn }
 
-  name                = each.value.name
-  location            = var.location
-  resource_group_name = var.resource_group_name
+  name                       = each.value.name
+  location                   = var.location
+  resource_group_name        = var.resource_group_name
   type                       = "IPsec"
   virtual_network_gateway_id = azurerm_virtual_network_gateway.vnetgw.id
   local_network_gateway_id   = azurerm_local_network_gateway.lnetgw[each.value.local_network_gateway_name].id
-  shared_key = each.value.shared_key
-  connection_mode = "Default"
-  connection_protocol = each.value.connection_protocol
-  enable_bgp = each.value.enable_bgp
+  shared_key                 = each.value.shared_key
+  connection_mode            = "Default"
+  connection_protocol        = each.value.connection_protocol
+  enable_bgp                 = each.value.enable_bgp
 
   dynamic "custom_bgp_addresses" {
-    for_each = each.value.custom_bgp_addresses != null ? [true] : []
+    for_each = each.value.custom_bgp_addresses != null ? [each.value.custom_bgp_addresses] : []
 
     content {
-      primary = each.value.custom_bgp_addresses.primary
-      secondary = each.value.custom_bgp_addresses.secondary
+      primary   = custom_bgp_addresses.value.primary
+      secondary = custom_bgp_addresses.value.secondary
     }
   }
 
   dynamic "ipsec_policy" {
-    for_each = each.value.ipsec_policy!= null ? [true] : []
+    for_each = each.value.ipsec_policy != null ? [each.value.ipsec_policy] : []
 
     content {
-      dh_group         = each.value.ipsec_policy.dh_group
-      ike_encryption   = each.value.ipsec_policy.ike_encryption
-      ike_integrity    = each.value.ipsec_policy.ike_integrity
-      ipsec_encryption = each.value.ipsec_policy.ipsec_encryption
-      ipsec_integrity  = each.value.ipsec_policy.ipsec_integrity
-      pfs_group        = each.value.ipsec_policy.pfs_group
-      sa_datasize      = each.value.ipsec_policy.sa_datasize
-      sa_lifetime      = each.value.ipsec_policy.sa_lifetime
+      dh_group         = ipsec_policy.value.dh_group
+      ike_encryption   = ipsec_policy.value.ike_encryption
+      ike_integrity    = ipsec_policy.value.ike_integrity
+      ipsec_encryption = ipsec_policy.value.ipsec_encryption
+      ipsec_integrity  = ipsec_policy.value.ipsec_integrity
+      pfs_group        = ipsec_policy.value.pfs_group
+      sa_datasize      = ipsec_policy.value.sa_datasize
+      sa_lifetime      = ipsec_policy.value.sa_lifetime
     }
   }
 }